+++ examples/01_basic.rs
@@ -24,5 +24,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }