+++ examples/02_serve_file.rs
@@ -36,5 +36,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }