+++ examples/03_routeing.rs
@@ -47,5 +47,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }