+++ examples/04_data.rs
@@ -61,5 +61,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }