+++ examples/09_cookie.rs
@@ -57,5 +57,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }