+++ examples/11_logging.rs
@@ -40,5 +40,5 @@ fn main() {
 
     // Start the server
     // This will block the current thread
-    server.start();
+    server.start().unwrap();
 }