+++ Changelog.md
@@ -1,6 +1,8 @@
 # 0.1.7*
 - Add Panic Message to Error Handel
 - Add http.rs to move raw http parseing out of server.rs
+- Start / Start Threaded returns Option
+- Add .unwrap to all server.starts in examples
 
 # 0.1.6
 - Add Example for Logging